User prompts for install

+++ install.sh
@@ -6,70 +6,92 @@
 #
 # zsh
 
-echo "[*] Installing zsh"
+read -p "Install zsh?" -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    echo "[*] Installing zsh"
 
-sudo apt install zsh -y
+    sudo apt install zsh -y
 
-echo "[*] Downloading oh-my-zsh"
+    echo "[*] Downloading oh-my-zsh"
 
-rm -rf ~/.oh-my-zsh
-git clone git://github.com/robbyrussell/oh-my-zsh.git ~/.oh-my-zsh
+    rm -rf ~/.oh-my-zsh
+    git clone git://github.com/robbyrussell/oh-my-zsh.git ~/.oh-my-zsh
 
-echo "[*] Linking zshrc"
+    echo "[*] Linking zshrc"
 
-ln -sf /home/kdasgupta/workspace/dotfiles/_zshrc ~/.zshrc
+    ln -sf /home/kdasgupta/workspace/dotfiles/_zshrc ~/.zshrc
+fi
 
 #
 # tmux
 
-echo "[*] Installing tmux"
+read -p "Install tmux?" -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    echo "[*] Installing tmux"
 
-sudo apt install tmux -y
+    sudo apt install tmux -y
 
-echo "[*] Linking tmux.conf"
+    echo "[*] Linking tmux.conf"
 
-ln -sf /home/kdasgupta/workspace/dotfiles/_tmux.conf ~/.tmux.conf
+    ln -sf /home/kdasgupta/workspace/dotfiles/_tmux.conf ~/.tmux.conf
+fi
 
 #
 # nvim
 
-echo "[*] nvim installer script..."
+read -p "Install nvim?" -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    echo "[*] nvim installer script..."
 
-(cd config/nvim && sh install.sh)
+    (cd config/nvim && sh install.sh)
+fi
 
 #
 # i3 + polybar
 # https://github.com/regolith-linux/speed-ricer
 
-echo "[*] Installing i3-gaps and polybar"
 
-sudo add-apt-repository -y ppa:kgilmer/speed-ricer
-sudo apt install i3-gaps polybar -y
+read -p "Install i3 polybar compton Xresources?" -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    echo "[*] Installing i3-gaps and polybar"
 
-echo "[*] Linking configs"
+    sudo add-apt-repository -y ppa:kgilmer/speed-ricer
+    sudo apt install i3-gaps polybar -y
 
-ln -sf /home/kdasgupta/workspace/dotfiles/config/i3 ~/.config/i3
-ln -sf /home/kdasgupta/workspace/dotfiles/config/polybar ~/.config/polybar
+    echo "[*] Linking configs"
 
-echo "[*] Installing compton"
-sudo apt install compton -y
+    ln -sf /home/kdasgupta/workspace/dotfiles/config/i3 ~/.config/i3
+    ln -sf /home/kdasgupta/workspace/dotfiles/config/polybar ~/.config/polybar
 
-echo "[*] Linking configs"
-ln -sf /home/kdasgupta/workspace/dotfiles/config/compton ~/.config/compton
+    echo "[*] Installing compton"
+    sudo apt install compton -y
 
-#
-# Xresources
+    echo "[*] Linking configs"
+    ln -sf /home/kdasgupta/workspace/dotfiles/config/compton ~/.config/compton
 
-echo "[*] Linking .Xresources"
+    #
+    # Xresources
 
-ln -sf /home/kdasgupta/workspace/dotfiles/_Xresources ~/.Xresources
+    echo "[*] Linking .Xresources"
+
+    ln -sf /home/kdasgupta/workspace/dotfiles/_Xresources ~/.Xresources
+
+fi
 
 #
 # gnupg
 
-echo "[*] Linking gnupg"
+# echo "[*] Linking gnupg"
 
-ln -sf /home/kdasgupta/workspace/dotfiles/_gnupg ~/.gnupg
+# ln -sf /home/kdasgupta/workspace/dotfiles/_gnupg ~/.gnupg
 
 ###################################################
